Extract `spec/layer.ts`
